Move diff formatting to different file
